working delete authorization

+++ controllers/poster.js
@@ -11,10 +11,19 @@ const router = express.Router()
 router.use((req, res, next) => {
     if(req.session.loggedIn) {
         next()
+        
         // logic for the profile button
     } else {
-        res.redirect('/user/login')
-    }
+
+        // res.redirect('/')
+            // get the data from the database
+            Poster.find({})
+            .then((posters) => {
+                res.render('index.ejs', {posters})
+                
+            })
+            .catch(err => console.log(err)) 
+}
 })
 
 // Poster Routes 
@@ -23,29 +32,36 @@ router.use((req, res, next) => {
 
 // INDEX route
 router.get('/', (req, res) => {
-
+    
     // get the data from the database
     Poster.find({})
     .then((posters) => {
-        res.render('index.ejs', {posters})
+        res.render('posters/index.ejs', {posters})
         console.log(posters)
     })
     .catch(err => console.log(err))
 })
+
+
 
 // Report route
 router.get('/report', (req, res) => {
     res.render('posters/report.ejs')
 })
 
+
+
 // NEW route
 router.get('/new', (req, res) => {
-    res.render('posters/new.ejs')
+    res.render('posters/new.ejs', {user: req.session.username})
 })
+
+
 
 // POST route
 router.post('/', (req, res) => {
- 
+    req.body.username = req.session.username
+
     // create new poster
     Poster.create(req.body, (err, createdPoster) => {
         res.redirect('/posters')
@@ -53,36 +69,6 @@ router.post('/', (req, res) => {
     })
     
 })
-
-
-
-
-
-
-//report
-
-// router.get('/:id/report', (req, res) => {
-//     //getting the poster from database
-//     console.log(req.params.id)
-//     Poster.findById(req.params.id, (err, foundPoster) => {
-//         res.render('posters/report.ejs', {
-//             poster: foundPoster,
-//             id: req.params.id
-//         })
-        
-//     })
-// })
-
-// router.put('/:id', (req, res) => {
-
-//     // req.body.status = req.body.status === "on" ? "Lost" : "Found"
-
-//     Poster.findByIdAndUpdate(req.params.id, req.body, {new: true}, (err, updatedPoster) => {
-//         res.redirect(`/posters`)
-//     })
-// })
-
-
 
 
 
@@ -99,12 +85,13 @@ router.get('/:id/edit', (req, res) => {
 
 router.put('/:id', (req, res) => {
 
-   
-
     Poster.findByIdAndUpdate(req.params.id, req.body, {new: true}, (err, updatedPoster) => {
         res.redirect(`/posters/${req.params.id}`)
     })
 })
+
+
+
 
 
 // SHOW route
@@ -118,14 +105,30 @@ router.get('/:id', (req, res) => {
  })
 
 
+
+
  // DELETE route 
  router.delete('/:id', (req, res) => {
-    //find by id and delete
-
-    Poster.findByIdAndDelete(req.params.id, (err, deletedPoster) => {
-        res.redirect('/posters')
+  
+    // find the poster by id and check if req.session.username is the owner of the poster
+    // only if true delete
+   Poster.findById(req.params.id)
+    .then((poster) => {
+        if(req.session.username === poster.username) {
+              //find by id and delete
+            Poster.findByIdAndDelete(req.params.id, (err, deletedPoster) => {
+               res.redirect('/posters')
+            })
+       } else {
+            res.send('you are not authorized to delete this post')
+       }
+        console.log(poster)
     })
+    .catch(err => console.log(err))
+ 
+    // console.log(poster)
  })
 
+ 
 
 module.exports = router
